requirements.txt
openpyxl
python-pptx
python-docx
google-genai
python-dotenv
PyMuPDF
pydantic>=2.0
